+++ webclient/src/views/DetailsView.vue
@@ -8,11 +8,11 @@ import DetailsFeedbackButton from "@/components/DetailsFeedbackButton.vue";
 import DetailsRoomfinderMap from "@/components/DetailsRoomfinderMap.vue";
 //import DetailsFeaturedSection from "@/components/DetailsFeaturedSection.vue";
 import { useI18n } from "vue-i18n";
-import { getLocalStorageWithExpiry, removeLocalStorage } from "@/utils/storage";
-import { copyCurrentLink, setDescription, setTitle } from "@/utils/common";
+import { getLocalStorageWithExpiry, removeLocalStorage } from "@/composables/storage";
+import { copyCurrentLink, setDescription, setTitle } from "@/composables/common";
 import { selectedMap, useDetailsStore } from "@/stores/details";
 import { nextTick, onMounted, ref, watch } from "vue";
-import { useFetch } from "@/utils/fetch";
+import { useFetch } from "@/composables/fetch";
 import { useRoute } from "vue-router";
 import router from "@/router";
 import type { components } from "@/api_types";
